worker/pyproject.toml
[project]
name = "netboxlabs-worker"
version = "0.0.1"  # Overwritten during the build process
description = "NetBox Labs, Worker backend for Orb Agent, part of NetBox Discovery"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    {name = "NetBox Labs", email = "support@netboxlabs.com" }
]
maintainers = [
    {name = "NetBox Labs", email = "support@netboxlabs.com" }
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
]

dependencies = [
    "APScheduler~=3.10",
    "croniter~=5.0",
    "fastapi~=0.115",
    "httpx~=0.27",
    "netboxlabs-diode-sdk~=0.4",
    "pydantic~=2.9",
    "uvicorn~=0.32",
]

[project.optional-dependencies]
dev = ["black", "check-manifest", "ruff"]
test = ["coverage", "pytest", "pytest-cov"]

[project.urls]
"Homepage" = "https://netboxlabs.com/"

[project.scripts]
worker = "worker.main:main"

[tool.setuptools]
packages = [
    "worker",
    "worker.policy",
]
package-data = {"worker" = ["**/*", "policy/**"]}
exclude-package-data = {worker = ["tests/*"]}

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.ruff]
line-length = 140

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["C", "D", "E", "F", "I", "R", "UP", "W"]
ignore = ["F401", "D203", "D212", "D400", "D401", "D404", "RET504"]